fix: moved LoginOptions
Related work item: GH-5
